+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ HEAD
   - add :ckey:`degradation.log_collect` to store the output of precollect phase in isolated logs
   - add working ``--compute-missing`` parameter to check group, which temporarily sets the precollection
   - add repetition of the time collector
+  - add predefined configuration templates
 
 0.14.2 (2018-05-15)
 -------------------